requirements.txt
feedparser==6.0.11
google-generativeai>=0.8.3
tweepy==4.14.0
python-dotenv==1.0.1
